Do not depend on strum_macros
The "strum" crate exposes the "derive" feature, which may be used to
be able to access the derive macros instead of depending on the
strum_derive crate directly.

This was actually found to be an issue we encountered in a project where
we depend on xee-interpreter and some other crates that pull in two
different versions of strum (transitively).

The project setup was like this:

    strum <-- indextree-macros <-- .. <-- xee-interpreter <-- our crate
    strum <-- xee-interpreter <-- our crate

In our setup, we use cargo-hakari for unification of dependencies for
faster builds.

The issue was not present here, because _before_ the update, the
"indextree-macros" dependency chain depended on another version of
"strum" than the other dependency chain.

Now when we ran cargo-update, the "strum" dependency matched in the
whole dependency tree, which causes cargo-hakari to be able to unify the
features.

This enabled the "derive" feature for the direct dependency on "strum"
for "xee-interpreter", which then failed the build because of the
duplicated import of the "EnumMessage" symbol, which was imported
automatically by rust as a derive macro, plus it was imported in
the codebase via `use strum_macros::EnumMessage`. This duplication
caused the build failure.

The thing that actually caused this issue was
the indextree-macros crate updating the "strum" dependency in a patch
release, which should never have happened.

This is a rather obscure issue, but it was fun to debug :-D

Co-authored-by: Marcel Müller <neikos@neikos.email>
Suggested-by: Marcel Müller <neikos@neikos.email>
Signed-off-by: Marcel Müller <neikos@neikos.email>
Signed-off-by: Matthias Beyer <mail@beyermatthias.de>